Move graphics related code to separate file
