+++ src/game.rs
@@ -1,2 +1,4 @@
 pub mod game;
+
 mod components;
+mod graphics;
